Fixed up more 64 bit warnings. Moved CSS Parser Generator into CSS build as build phase instead of separate target.

+++ src/pagingscrollview/src/NIPagingScrollView.m
@@ -346,7 +346,7 @@ const CGFloat NIPagingScrollViewDefaultPageMargin = 10;
     }
 
     // Add missing pages.
-    for (int pageIndex = rangeOfVisiblePages.location;
+    for (NSUInteger pageIndex = rangeOfVisiblePages.location;
          pageIndex < (NSInteger)NSMaxRange(rangeOfVisiblePages); ++pageIndex) {
       if (![self isDisplayingPageForIndex:pageIndex]) {
         [self displayPageAtIndex:pageIndex];
